Deprecate MessageRoutingInformation
Even though MessageRoutingInformation and MemberCapabilities resemble
each other quite a lot, we'll move to the newer variant. Partially
because it's a move from implementation to interface, but also due to
the deprecation of the SpringCloudHttpBackupCommandRouter which is the
sole user of the MessageRoutingInformation. MemberCapabilities on the
other hand is core to the capability discovery mode implementations

#23