Deprecate the SpringCloudHttpBackupCommandRouter
The entire SpringCloudHttpBackupCommandRouter should be deprecated, but
should also still be usable. Towards that end, all overridden methods
have been removed from the CommandRouter, deprecating the
getLocalMessageRoutingInformation further more and defaulting it to a
none-usable MessageRoutingInformation instance. The builder of the
CommandRouter should default to a RestCapabilityDiscoveryMode instance
which is ignore-listing instances, as that was the original approach of
this router. Any REST specific annotations should also be removed from
this Router impl. The javadoc should correctly reflect the left
over intent of this router. Lastly, the existing test cases should be
validated for the worth and copied over to the respective replacement
discovery mode implementations (if applicable).

#23