Fine tune the IgnoreListingCapabilityDiscoveryMode
- Store the entire ServiceInstance instead of just the identifier. Doing
 so will uncover reuse of id's, with different host/port combinations
- To match ServiceInstances, reuse the custom equals method which was
present in the SpringCloudCommandRouter. Doing so ensure we cover for
scenarios where the ServiceInstance does not implement equals
- Clean up the javadoc accordingly
- Remove the cleanIgnoredInstanceSet operation. Any conceivable
solutions in this space reflect the intent of issue #6. To not pull in
that issue too, that logic should be removed to that issue for a
following release. Furthermore, the old approach of clean up, which
based itself on the entire set of discoverable ServiceInstances no
longer works, as we fetch capabilities per instances, not all of them at
once

#1647