Add local instance in absence of any found instances
If the DiscoveryClient does not return any ServiceInstances, an update
of the consistent hash would mean nobody can handle anything. In reality
 though, the node itself can typically still handle something. Hence the
 localServiceInstance should be included if the instances set is empty.
Note that not doing this is an issue when testing with a Eureka
discovery service. Right after the InstanceRegisteredEvent we clear out
the local instance to correctly reflect the URI it should at that point
at least have. The DiscoveryClient at that stage however still doesn't
know anything, not even the local registration. Hence the consistent
hash could incorrectly be cleared out in such a scenario, which is thus
covered by adding it if no services are found.

#23